# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing airport tools",
        "title": "Airport Tools Vision API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/transaction": {
            "get": {
                "description": "Возвращает список транзакций QA.\u003cbr\u003e Можно фильтровать по статусу с помощью query-параметра `status`.\u003cbr\u003e Допустимое значение: 'qa' вернёт только транзакции, требующие проверки QA.\u003cbr\u003e Каждая транзакция содержит минимальные данные: ID, инженера, дату создания, текущий статус.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Список транзакций",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу транзакции",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ListTransactionsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/transaction/check": {
            "post": {
                "description": "Принимает табельный номер инженера и фотографию инструментов в формате base64.\u003cbr\u003e Сервис анализирует изображение, сопоставляет инструменты с ожидаемым набором и возвращает: \u003cbr\u003e\u003cbr\u003e• URL обработанного изображения \u003cbr\u003e• четыре массива: \u003cbr\u003e1) access_tools — инструменты, прошедшие автоматическую проверку\u003cbr\u003e2) manual_check_tools — инструменты, требующие ручной проверки \u003cbr\u003e3) unknown_tools — инструменты, отсутствующие в ожидаемом наборе \u003cbr\u003e4) missing_tools — инструменты, отсутствующие на фотографии, но ожидаемые\u003cbr\u003e• transaction_type - тип транзакции(Checkin - Сдача/Checkout - Выдача)\u003cbr\u003e• status - статус транзакции(OPEN - открыта, CLOSED - закрыта, QA VERIFICATION - QA проверка)\u003cbr\u003e\u003cbr\u003e Если 4 или более инструментов не попали в access_tools или за 3 попытки сканирования транзакция не закрылась, устанавливается флаг \"QA ПРОВЕРКА\" (QA VERIFICATION). \u003cbr\u003e\u003cbr\u003eЭндпоинт используется как для выдачи инструментов инженеру, так и для их последующей сдачи.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Операция выдачи/сдачи инструментов",
                "parameters": [
                    {
                        "description": "Запрос на выдачу или сдачу инструментов",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CheckReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CheckRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/transaction/{transaction_id}/verification": {
            "get": {
                "description": "Получить информацию о проблемной транзакции.\u003cbr\u003eОткрывается экран сверки:\u003cbr\u003e\u003cbr\u003e • Фотография инструментов (полноразмерное изображение)\u003cbr\u003e • access_tools — инструменты, прошедшие автоматическую проверку\u003cbr\u003e • Список проблемных инструментов с пояснениями, сгруппированных по категориям:\u003cbr\u003e \u0026nbsp;\u0026nbsp;2) manual_check_tools — инструменты, требующие ручной проверки\u003cbr\u003e \u0026nbsp;\u0026nbsp;3) unknown_tools — инструменты, не входящие в ожидаемый набор\u003cbr\u003e \u0026nbsp;\u0026nbsp;4) missing_tools — инструменты, отсутствующие на фото, но ожидаемые\u003cbr\u003e\u003cbr\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Получение информации о проблемной транзакции",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор транзакции",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные о транзакции",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetQAVerificationRes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.VerificationRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "После авторизации сотрудника QA выбирает из списка транзакцию.\u003cbr\u003eОткрывается экран сверки:\u003cbr\u003e\u003cbr\u003e • Фотография инструментов (полноразмерное изображение)\u003cbr\u003e • access_tools — инструменты, прошедшие автоматическую проверку\u003cbr\u003e • Список проблемных инструментов с пояснениями, сгруппированных по категориям:\u003cbr\u003e \u0026nbsp;\u0026nbsp;1) ManualCheckTools — инструменты, требующие ручной проверки\u003cbr\u003e \u0026nbsp;\u0026nbsp;2) UnknownTools — инструменты, не входящие в ожидаемый набор\u003cbr\u003e \u0026nbsp;\u0026nbsp;3) MissingTools — инструменты, отсутствующие на фото, но ожидаемые\u003cbr\u003e\u003cbr\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "QA-проверка и завершение транзакции",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор транзакции",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные завершения QA-проверки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.VerificationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.VerificationRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "description": "Вход в систему по табельному номеру сотрудника.\u003cbr\u003e После успешного входа пользователь перенаправляется:\u003cbr\u003e • инженеру — на экран загрузки фотографии инструментов;\u003cbr\u003e • QA — на экран проверки незавершённых транзакций.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.LoginRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/register": {
            "post": {
                "description": "Регистрация сотрудника в системе.\u003cbr\u003e Необходимые данные: табельный номер, ФИО и роль (например, \"Инженер\" или \"QA\").\u003cbr\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Регистрация сотрудника в системе",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1.RegisterRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/roles": {
            "get": {
                "description": "Возвращает список всех возможных ролей пользователей в системе.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получить список ролей",
                "responses": {
                    "200": {
                        "description": "Список ролей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.GetRolesRes"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Role": {
            "type": "string",
            "enum": [
                "Engineer",
                "Quality Auditor"
            ],
            "x-enum-comments": {
                "Engineer": "Авиатехник / Инженер",
                "QualityAuditor": "Специалист службы качества / аудит"
            },
            "x-enum-descriptions": [
                "Авиатехник / Инженер",
                "Специалист службы качества / аудит"
            ],
            "x-enum-varnames": [
                "Engineer",
                "QualityAuditor"
            ]
        },
        "v1.CheckReq": {
            "type": "object",
            "required": [
                "data",
                "employee_id"
            ],
            "properties": {
                "data": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "string"
                },
                "tool_set_id": {
                    "type": "integer"
                }
            }
        },
        "v1.CheckRes": {
            "type": "object",
            "properties": {
                "access_tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.RecognizedToolDTO"
                    }
                },
                "debug_image_url": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "problematic_tools": {
                    "$ref": "#/definitions/v1.ProblematicTools"
                },
                "status": {
                    "type": "string"
                },
                "transaction_type": {
                    "type": "string"
                }
            }
        },
        "v1.GetQAVerificationRes": {
            "type": "object",
            "properties": {
                "access_tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.RecognizedToolDTO"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "problematic_tools": {
                    "$ref": "#/definitions/v1.ProblematicTools"
                },
                "tool_set_id": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/v1.UserDto"
                }
            }
        },
        "v1.GetRolesRes": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Role"
                    }
                }
            }
        },
        "v1.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "v1.ListTransactionsRes": {
            "type": "object",
            "properties": {
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.TransactionDTO"
                    }
                }
            }
        },
        "v1.LoginReq": {
            "type": "object",
            "required": [
                "employee_id"
            ],
            "properties": {
                "employee_id": {
                    "type": "string"
                }
            }
        },
        "v1.LoginRes": {
            "type": "object",
            "properties": {
                "role": {
                    "$ref": "#/definitions/domain.Role"
                }
            }
        },
        "v1.ProblematicTools": {
            "type": "object",
            "properties": {
                "manual_check_tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.RecognizedToolDTO"
                    }
                },
                "missing_tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.ToolTypeDTO"
                    }
                },
                "unknown_tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.RecognizedToolDTO"
                    }
                }
            }
        },
        "v1.RecognizedToolDTO": {
            "type": "object",
            "properties": {
                "bbox": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "confidence": {
                    "type": "number"
                },
                "tool_type_id": {
                    "type": "integer"
                }
            }
        },
        "v1.RegisterReq": {
            "type": "object",
            "required": [
                "employee_id",
                "full_name",
                "role"
            ],
            "properties": {
                "employee_id": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/domain.Role"
                }
            }
        },
        "v1.RegisterRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "v1.ToolTypeDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "part_number": {
                    "type": "string"
                }
            }
        },
        "v1.TransactionDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tool_set_id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/v1.UserDto"
                }
            }
        },
        "v1.UserDto": {
            "type": "object",
            "properties": {
                "employee_id": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                }
            }
        },
        "v1.VerificationReq": {
            "type": "object",
            "required": [
                "qa_employee_id"
            ],
            "properties": {
                "notes": {
                    "type": "string"
                },
                "qa_employee_id": {
                    "type": "string"
                }
            }
        },
        "v1.VerificationRes": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Время завершения проверки",
                    "type": "string"
                },
                "status": {
                    "description": "Новый статус",
                    "type": "string"
                },
                "transaction_id": {
                    "description": "ID транзакции, которую QA завершил",
                    "type": "integer"
                },
                "verified_by": {
                    "description": "Табельный номер или имя QA",
                    "type": "string"
                }
            }
        }
    }
}